+++ src/index.ts
@@ -1,3 +1,4 @@
-import LayerManager from './LayerManager';
-
-export default LayerManager;
+export { default as LayerManager, default } from './LayerManager';
+export { LayerAnalyzer } from './LayerAnalyzer';
+export type { ILayerManager, LayerConfig, Source, LayerManagerOptions } from './interfaces';
+export type { AnalyzerReport, FrameStats, SourceLoadTime, LayerTimeStat } from './LayerAnalyzer';