+++ src/utils.ts
@@ -1,16 +1,16 @@
-import { LayerConfig } from 'interfaces';
+import { LayerConfig } from './interfaces';
 import { Layer } from 'mapbox-gl';
 
 export const extendLayerWithConfig = (layer: Layer, config: LayerConfig): Layer => {
+  const cloned: Layer = { ...layer };
   if (config.filter) {
-    layer.filter = ['all', config.filter];
+    cloned.filter = ['all', config.filter];
   }
   if (config.layout) {
-    layer.layout = config.layout;
+    cloned.layout = config.layout;
   }
   if (config.paint) {
-    layer.paint = config.paint;
+    cloned.paint = config.paint;
   }
-
-  return layer;
+  return cloned;
 };